Refatorada estrutura do projeto
