+++ backend/Authenticate/Dtos/AuthUserReturnDto.cs
@@ -1,0 +1,9 @@
+using backend.Shared.Dtos;
+
+namespace backend.Authenticate.Dtos
+{
+  public class AuthUserReturnDto : ReturnDto
+  {
+    public string JWT { get; set; }
+  }
+}
