+++ backend/Authenticate/Dtos/VerifyTokenReturnDto.cs
@@ -1,0 +1,9 @@
+using backend.Shared.Dtos;
+
+namespace backend.Authenticate.Dtos
+{
+  public class VerifyTokenReturnDto : ReturnDto
+  {
+    public string Token { get; set; }
+  }
+}
