+++ backend/Shared/Dtos/ReturnDto.cs
@@ -1,0 +1,9 @@
+using backend.Messages;
+
+namespace backend.Shared.Dtos
+{
+  public class ReturnDto
+  {
+    public Message Message { get; set; }
+  }
+}
