+++ backend/Shared/Users/Dtos/GetUserByIdReturnDto.cs
@@ -1,0 +1,7 @@
+namespace backend.Shared.Dtos
+{
+  public class GetUserByIdReturnDto : ReturnDto
+  {
+    public UserReadDto User { get; set; }
+  }
+}
