+++ backend/UserSettings/Dtos/UpdatePasswordDto.cs
@@ -1,0 +1,8 @@
+namespace backend.UserSettings.Dtos
+{
+  public class UpdatePasswordDto
+  {
+    public string ActualPassword { get; set; }
+    public string NewPassword { get; set; }
+  }
+}
