--- backend/dtos/AuthUserReturnDto.cs
@@ -1,7 +1,0 @@
-namespace backend.dtos
-{
-  public class AuthUserReturnDto : ReturnDto
-  {
-    public string JWT { get; set; }
-  }
-}
