--- backend/dtos/GetUserByIdReturnDto.cs
@@ -1,9 +1,0 @@
-using backend.models;
-
-namespace backend.dtos
-{
-  public class GetUserByIdReturnDto : ReturnDto
-  {
-    public UserReadDto User { get; set; }
-  }
-}
