--- backend/dtos/ReturnDto.cs
@@ -1,9 +1,0 @@
-using backend.Messages;
-
-namespace backend.dtos
-{
-  public class ReturnDto
-  {
-    public Message Message { get; set; }
-  }
-}
