--- backend/dtos/UpdatePasswordDto.cs
@@ -1,8 +1,0 @@
-namespace backend.dtos
-{
-  public class UpdatePasswordDto
-  {
-    public string ActualPassword { get; set; }
-    public string NewPassword { get; set; }
-  }
-}
