--- backend/dtos/UserCreateDtos.cs
@@ -1,22 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace backend.models
-{
-  public class UserCreateDto
-  {
-
-    [Required]
-    [MaxLength(30)]
-    public string name { get; set; } = "";
-
-    
-    [Required]
-    [MaxLength(256)]
-    public string email { get; set; } = "";
-
-    [Required]
-    [MaxLength(250)]
-    public string password { get; set; } = "";
-
-  }
-}
