--- backend/dtos/VerifyTokenReturnDto.cs
@@ -1,7 +1,0 @@
-namespace backend.dtos
-{
-  public class VerifyTokenReturnDto : ReturnDto
-  {
-    public string Token { get; set; }
-  }
-}
