+++ src/app/home/modules/user-config/pages/profile/profile.service.ts
@@ -1,12 +1,13 @@
+import { UserConfigProxyService } from '../../../../../shared/proxys/user-config-proxys/user-config.proxy.service';
 import { Router } from '@angular/router';
 import { Injectable } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
 
-import { take, Subscription } from 'rxjs';
+import { take, } from 'rxjs';
 
 import { UserHandlerService } from '../../../../../shared/handlers/user-handler.service';
 import User from '../../../../../shared/support/interfaces/user.interface';
-import { UserCrudProxysService } from '../../../../../shared/proxys/userCrudProxys/user-crud-proxys.service';
+
 import IProfileService from './IProfile.service.interface';
 import ResponseDto from 'src/app/shared/support/classes/responseDto';
 
@@ -16,12 +17,12 @@ import ResponseDto from 'src/app/shared/support/classes/responseDto';
 export class ProfileService implements IProfileService {
 
   constructor(
-    private userCrudProxysService: UserCrudProxysService,
+    private userConfigProxyService: UserConfigProxyService,
     private userHandlerService: UserHandlerService,
     private router:Router
   ) {}
   updateProfilePreferences(userId: string, user: User, callback?: Function): void {
-    this.userCrudProxysService.updateUserRequest(userId, user)
+    this.userConfigProxyService.updateUserRequest(userId, user)
       .pipe(
         take(1)
       )
