+++ src/app/shared/handlers/user-handler.service.ts
@@ -3,7 +3,7 @@ import { Injectable, EventEmitter } from '@angular/core';
 
 import { Observable, take } from 'rxjs';
 
-import { UserCrudProxysService } from '../proxys/userCrudProxys/user-crud-proxys.service';
+import { UserCrudProxysService } from '../proxys/userCrudProxys/user-crud.proxys.service';
 import IUserHandlerService from './IUser-handler.service.interface';
 
 @Injectable({
