+++ src/app/shared/proxys/userCrudProxys/IUser-Crud.proxys.service.interface.ts
@@ -5,6 +5,4 @@ import UserInput from "../../support/interfaces/userInput.interface";
 export default interface IUserCrudProxysService {
     createNewUserRequest: (user: UserInput) => Observable<any>
     getUserByIdRequest: (id: string) => Observable<any>
-    updateUserRequest: (userId: string, user: User) => Observable<any>
-
 }
